token functions in util folder
