+++ api/auth/util/signToken.js
@@ -1,0 +1,19 @@
+const jwt = require('jsonwebtoken');
+
+module.exports = user => {
+	const payload = {
+		uid: user.id,
+		username: user.username,
+		isAdmin: user.isAdmin,
+		superUser: user.superUser,
+		isLocked: user.isLocked
+	};
+
+	const secret = process.env.JWT_SECRET + user.password;
+
+	const options = {
+		expiresIn: '24h'
+	};
+
+	return jwt.sign(payload, secret, options);
+};
